requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
Flask-Admin==1.6.1
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
